Update Module 1.2 Assignment.md
In the Assignment is says to create a directory with using ```mkdir``` but is not specifying the name of the directory. Someone new to the Command line will get lost when they see an error message that says the following ```usage: mkdir [-pv] [-m mode] directory_name ...``` . There are people that are literal and are neurodiverse and get loss or over whelm when doing something and it doesn't work. Just a quick fix on adding the name the directory will fix it.